pyproject.toml
[project]
name = "daily-report-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.0.4",
    "google-genai>=1.36.0",
    "psycopg>=3.2.10",
    "psycopg-binary>=3.2.10",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.43",
]

